pyproject.toml
[tool.poetry]
name = "challenge-lexmax"
version = "0.1.0"
description = ""
authors = ["Andres Armas <andresarmas97@hotmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0.4"
djangorestframework = "^3.13.1"
drf-spectacular = "^0.22.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
